incremental changes.  unsure why bearer tokens no longer function.

+++ spring-boot-monolith/src/main/java/com/example/springbootmonolith/repository/UserProfileRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserProfileRepository extends CrudRepository<UserProfile, Long> {
+public interface UserProfileRepository extends CrudRepository<UserProfile, Integer> {
 
     @Query("from UserProfile up left join User u on u.username = ?1 and up.id = u.userProfile.id")
     public UserProfile findProfileByUsername(String username);
